+++ fixtures/package/vocab.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  devLanguage: 'en',
+  languages: [{ name: 'en' }, { name: 'fr' }],
+  ignore: ['**/dist/**'],
+};