+++ fixtures/server/src/server.tsx
@@ -1,5 +1,4 @@
 /* eslint-disable no-console */
-import React from 'react';
 import { renderToString } from 'react-dom/server';
 import { ChunkExtractor } from '@loadable/server';
 import path from 'path';