+++ fixtures/server/webpack.config.js
@@ -31,7 +31,7 @@ module.exports = () => [
                 presets: [
                   ['@babel/preset-env', { modules: false }],
                   '@babel/preset-typescript',
-                  '@babel/preset-react',
+                  ['@babel/preset-react', { runtime: 'automatic' }],
                 ],
               },
             },
@@ -73,7 +73,7 @@ module.exports = () => [
                 presets: [
                   ['@babel/preset-env', { modules: false }],
                   '@babel/preset-typescript',
-                  '@babel/preset-react',
+                  ['@babel/preset-react', { runtime: 'automatic' }],
                 ],
               },
             },