+++ packages/core/src/utils.ts
@@ -16,6 +16,8 @@ export type Fallback = 'none' | 'valid' | 'all';
 
 const globAnyPathWithOptionalPrefix = '**/?(*)';
 
+export const compiledVocabFileFilter = /\.vocab[\\/]index\.(?:ts|js|cjs|mjs)$/;
+
 export function isDevLanguageFile(filePath: string) {
   return (
     filePath.endsWith(`/${devTranslationFileName}`) ||
@@ -36,18 +38,6 @@ export function isTranslationDirectory(
   return filePath.endsWith(translationsDirectorySuffix);
 }
 
-export function getTranslationFolderGlob({
-  translationsDirectorySuffix = defaultTranslationDirSuffix,
-}: {
-  translationsDirectorySuffix?: string;
-}) {
-  const result = `${globAnyPathWithOptionalPrefix}${translationsDirectorySuffix}`;
-
-  trace('getTranslationFolderGlob', result);
-
-  return result;
-}
-
 export function getDevTranslationFileGlob({
   translationsDirectorySuffix = defaultTranslationDirSuffix,
 }: {
@@ -66,7 +56,7 @@ export function getAltTranslationFileGlob(config: UserConfig) {
     altLanguages.length === 1 ? altLanguages[0] : `{${altLanguages.join(',')}}`;
 
   const { translationsDirectorySuffix = defaultTranslationDirSuffix } = config;
-  const result = `**/*${translationsDirectorySuffix}/${langMatch}.translations.json`;
+  const result = `${globAnyPathWithOptionalPrefix}${translationsDirectorySuffix}/${langMatch}.${devTranslationFileName}`;
 
   trace('getAltTranslationFileGlob', result);