+++ packages/vite/src/consts.ts
@@ -1,3 +1,8 @@
-export const compiledVocabFileFilter = /\.vocab[\\/]index\.(?:ts|js|cjs|mjs)$/;
+import { compiledVocabFileFilter as _compiledVocabFileFilter } from '@vocab/core';
+
+/**
+ * @deprecated Import from `@vocab/core` instead
+ */
+export const compiledVocabFileFilter = _compiledVocabFileFilter;
 export const virtualModuleId = 'virtual:vocab';
 export const sourceQueryKey = '?source=';