+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,88 @@
 # @vocab/cli
 
+## 2.1.12
+
+### Patch Changes
+
+- [#377](https://github.com/seek-oss/vocab/pull/377) [`12d07bc`](https://github.com/seek-oss/vocab/commit/12d07bca3de2e0ead5a2f066da00c4c4a981baae) Thanks [@jahredhope](https://github.com/jahredhope)! - Fix missing --branch option as string for push and pull commands.
+
+  Since v2.1.11, CLI commands to push and pull translations only accepted `--branch` as a boolean flag due to a bug. `--branch` can now be used with a string as intended.
+
+  **Example Usage**
+
+  ```sh
+  vocab push --branch my-branch
+  vocab pull --branch my-branch
+  ```
+
+  Fixes error "error: too many arguments for 'push'. Expected 0 arguments but got 1." and "error: too many arguments for 'pull'. Expected 0 arguments but got 1."
+
+## 2.1.11
+
+### Patch Changes
+
+- [#372](https://github.com/seek-oss/vocab/pull/372) [`9fd2e34`](https://github.com/seek-oss/vocab/commit/9fd2e344b61c666ac186e2478765c4d4d6a27f2f) Thanks [@askoufis](https://github.com/askoufis)! - Replace `yargs` dependency with `commander`.
+
+## 2.1.10
+
+### Patch Changes
+
+- [#361](https://github.com/seek-oss/vocab/pull/361) [`4d2604f`](https://github.com/seek-oss/vocab/commit/4d2604f4c093241b3a52f75b1a113202a1b6ab64) Thanks [@askoufis](https://github.com/askoufis)! - Remove `env-ci` dependency
+
+- Updated dependencies [[`f4ad654`](https://github.com/seek-oss/vocab/commit/f4ad654d926d3a7cdf6a477bc609a804ac4bb37c)]:
+  - @vocab/core@1.7.6
+  - @vocab/phrase@2.1.10
+
+## 2.1.9
+
+### Patch Changes
+
+- Updated dependencies [[`5f5bdb5`](https://github.com/seek-oss/vocab/commit/5f5bdb5e323c0121491d923dcaae961c5ac3fe81)]:
+  - @vocab/core@1.7.5
+  - @vocab/phrase@2.1.9
+
+## 2.1.8
+
+### Patch Changes
+
+- Updated dependencies [[`710649f`](https://github.com/seek-oss/vocab/commit/710649fae726227701f51410d6050fd0c17d6e3f)]:
+  - @vocab/core@1.7.4
+  - @vocab/phrase@2.1.8
+
+## 2.1.7
+
+### Patch Changes
+
+- [#359](https://github.com/seek-oss/vocab/pull/359) [`478c86e`](https://github.com/seek-oss/vocab/commit/478c86eedc99e461540f9b7fdfcabc6da27099c7) Thanks [@askoufis](https://github.com/askoufis)! - Remove unused `fast-glob` dependency
+
+- Updated dependencies [[`478c86e`](https://github.com/seek-oss/vocab/commit/478c86eedc99e461540f9b7fdfcabc6da27099c7), [`478c86e`](https://github.com/seek-oss/vocab/commit/478c86eedc99e461540f9b7fdfcabc6da27099c7)]:
+  - @vocab/core@1.7.3
+  - @vocab/phrase@2.1.7
+
+## 2.1.6
+
+### Patch Changes
+
+- Updated dependencies [[`6be1c6b`](https://github.com/seek-oss/vocab/commit/6be1c6b027a8f9471c1c9858c18d2750c2f33427)]:
+  - @vocab/core@1.7.2
+  - @vocab/phrase@2.1.6
+
+## 2.1.5
+
+### Patch Changes
+
+- Updated dependencies [[`e9d389b`](https://github.com/seek-oss/vocab/commit/e9d389b99aa8c6a93ad7464d5e6ea65e42851825)]:
+  - @vocab/core@1.7.1
+  - @vocab/phrase@2.1.5
+
+## 2.1.4
+
+### Patch Changes
+
+- Updated dependencies [[`a5e9d25`](https://github.com/seek-oss/vocab/commit/a5e9d256be271a47457213cb4eee5499af629116), [`a5e9d25`](https://github.com/seek-oss/vocab/commit/a5e9d256be271a47457213cb4eee5499af629116)]:
+  - @vocab/core@1.7.0
+  - @vocab/phrase@2.1.4
+
 ## 2.1.3
 
 ### Patch Changes