+++ packages/react/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @vocab/react
 
+## 1.1.22
+
+### Patch Changes
+
+- Updated dependencies [[`f4ad654`](https://github.com/seek-oss/vocab/commit/f4ad654d926d3a7cdf6a477bc609a804ac4bb37c)]:
+  - @vocab/core@1.7.6
+
+## 1.1.21
+
+### Patch Changes
+
+- Updated dependencies [[`5f5bdb5`](https://github.com/seek-oss/vocab/commit/5f5bdb5e323c0121491d923dcaae961c5ac3fe81)]:
+  - @vocab/core@1.7.5
+
+## 1.1.20
+
+### Patch Changes
+
+- Updated dependencies [[`710649f`](https://github.com/seek-oss/vocab/commit/710649fae726227701f51410d6050fd0c17d6e3f)]:
+  - @vocab/core@1.7.4
+
+## 1.1.19
+
+### Patch Changes
+
+- Updated dependencies [[`478c86e`](https://github.com/seek-oss/vocab/commit/478c86eedc99e461540f9b7fdfcabc6da27099c7), [`478c86e`](https://github.com/seek-oss/vocab/commit/478c86eedc99e461540f9b7fdfcabc6da27099c7)]:
+  - @vocab/core@1.7.3
+
+## 1.1.18
+
+### Patch Changes
+
+- Updated dependencies [[`6be1c6b`](https://github.com/seek-oss/vocab/commit/6be1c6b027a8f9471c1c9858c18d2750c2f33427)]:
+  - @vocab/core@1.7.2
+
+## 1.1.17
+
+### Patch Changes
+
+- Updated dependencies [[`e9d389b`](https://github.com/seek-oss/vocab/commit/e9d389b99aa8c6a93ad7464d5e6ea65e42851825)]:
+  - @vocab/core@1.7.1
+
+## 1.1.16
+
+### Patch Changes
+
+- Updated dependencies [[`a5e9d25`](https://github.com/seek-oss/vocab/commit/a5e9d256be271a47457213cb4eee5499af629116), [`a5e9d25`](https://github.com/seek-oss/vocab/commit/a5e9d256be271a47457213cb4eee5499af629116)]:
+  - @vocab/core@1.7.0
+
 ## 1.1.15
 
 ### Patch Changes