+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,72 @@
 # @vocab/vite
 
+## 0.3.9
+
+### Patch Changes
+
+- [#374](https://github.com/seek-oss/vocab/pull/374) [`acf25d2`](https://github.com/seek-oss/vocab/commit/acf25d2462fc75b52017425294be16cb62fdf6ee) Thanks [@renovate](https://github.com/apps/renovate)! - Update `es-module-lexer` dependency to `^2.0.0`
+
+## 0.3.8
+
+### Patch Changes
+
+- Updated dependencies [[`f4ad654`](https://github.com/seek-oss/vocab/commit/f4ad654d926d3a7cdf6a477bc609a804ac4bb37c)]:
+  - @vocab/core@1.7.6
+
+## 0.3.7
+
+### Patch Changes
+
+- Updated dependencies [[`5f5bdb5`](https://github.com/seek-oss/vocab/commit/5f5bdb5e323c0121491d923dcaae961c5ac3fe81)]:
+  - @vocab/core@1.7.5
+
+## 0.3.6
+
+### Patch Changes
+
+- Updated dependencies [[`710649f`](https://github.com/seek-oss/vocab/commit/710649fae726227701f51410d6050fd0c17d6e3f)]:
+  - @vocab/core@1.7.4
+
+## 0.3.5
+
+### Patch Changes
+
+- Updated dependencies [[`478c86e`](https://github.com/seek-oss/vocab/commit/478c86eedc99e461540f9b7fdfcabc6da27099c7), [`478c86e`](https://github.com/seek-oss/vocab/commit/478c86eedc99e461540f9b7fdfcabc6da27099c7)]:
+  - @vocab/core@1.7.3
+
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies [[`6be1c6b`](https://github.com/seek-oss/vocab/commit/6be1c6b027a8f9471c1c9858c18d2750c2f33427)]:
+  - @vocab/core@1.7.2
+
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies [[`e9d389b`](https://github.com/seek-oss/vocab/commit/e9d389b99aa8c6a93ad7464d5e6ea65e42851825)]:
+  - @vocab/core@1.7.1
+
+## 0.3.2
+
+### Patch Changes
+
+- [#351](https://github.com/seek-oss/vocab/pull/351) [`a5e9d25`](https://github.com/seek-oss/vocab/commit/a5e9d256be271a47457213cb4eee5499af629116) Thanks [@askoufis](https://github.com/askoufis)! - Deprecate `compiledVocabFileFilter`
+
+  `compiledVocabFileFilter` has been deprecated. Please import it from `@vocab/core` instead.
+
+  **MIGRATION GUIDE**:
+
+  ```diff
+  -import { compiledVocabFileFilter } from '@vocab/webpack';
+  -import { compiledVocabFileFilter } from '@vocab/vite';
+  +import { compiledVocabFileFilter } from '@vocab/core';
+  ```
+
+- Updated dependencies [[`a5e9d25`](https://github.com/seek-oss/vocab/commit/a5e9d256be271a47457213cb4eee5499af629116), [`a5e9d25`](https://github.com/seek-oss/vocab/commit/a5e9d256be271a47457213cb4eee5499af629116)]:
+  - @vocab/core@1.7.0
+
 ## 0.3.1
 
 ### Patch Changes